move configItems.js to config dir
