+++ server/lib/config/index.js
@@ -1,7 +1,7 @@
 const fs = require('fs')
 const path = require('path')
 const minimist = require('minimist')
-const definitions = require('../../resources/configItems')
+const definitions = require('./configItems')
 const fromDb = require('./fromDb')
 const fromDefault = require('./fromDefault')
 const fromEnv = require('./fromEnv')
